Adopt hashicorp/go-multierror.
The primary purpose is to allow the caller to unbundle a MultiError
into original errors and inspect each of them. It has technically
been possible, but the struct has been defined in an internal package.
It might be enough to move the package out of internal, but I thought
it would be nicer to use a popular package rather than our own so the
caller can handle bundled errors in a uniformed way.

The benefit will be realized after a subsequent change, where I plan
to define a new struct representing an error from Packager.Run.