Update regexes to support future SXG versions (#55)
This is a subset of the full set of allowed param values, defined to be token / quoted-string by https://tools.ietf.org/html/rfc7231#section-3.1.1.1, but attempts balance flexibility with feasibility. (An accurate definition would likely be long/complicated enough to result in more errors.)

Uses a PCRE syntax that was tested on nginx on Linux.