Put a size limit on the resource cache. (#79)
Change the default resource cache for webpkgserver to be size-bound (200
entries as a reasonable-seeming default), with options to make it unbounded or
to disable it.

This required changing Packager.RunForRequest to return the Resource it
creates, rather than having the caller rely on it being present as a side
effect in the ResourceCache. Otherwise, a cache eviction policy would be
incompatible with parallel calls to server.Handler.handleDocImpl (short of a
rather large critical section that would make it mostly not parallel).